Refactor item to project component
